put genre preferences as its own component
