+++ frontend/src/components/SignupPersonalize/SignupPersonalize.js
@@ -1,4 +1,3 @@
-import Select from 'react-select';
 import React, { useState, useEffect } from "react";
 import './SignupPersonalize.css'
 import { toast, ToastContainer } from 'react-toastify';
@@ -7,44 +6,8 @@ import { useNavigate } from 'react-router-dom';
 import SearchBooks from '../SearchBooks/SearchBooks';
 import axios from 'axios';
 import { set } from 'mongoose';
+import GenrePreferences from '../GenrePreferences/GenrePreferences';
 
-const genreOptions = [
-    { label: 'fiction', value: 'fiction' },
-    { label: 'mystery', value: 'mystery' },
-    { label: 'romance', value: 'romance' },
-    { label: 'science fiction', value: 'science fiction' },
-    { label: 'fantasy', value: 'fantasy' },
-    { label: 'thriller', value: 'thriller' },
-    { label: 'horror', value: 'horror' },
-    { label: 'historical fiction', value: 'historical fiction' },
-    { label: 'non-fiction', value: 'non-fiction' },
-    { label: 'biography', value: 'biography' },
-    { label: 'autobiography', value: 'autobiography' },
-    { label: 'memoir', value: 'memoir' },
-    { label: 'self-help', value: 'self-help' },
-    { label: 'young adult', value: 'young adult' },
-    { label: "children's literature", value: "children's literature" },
-    { label: 'poetry', value: 'poetry' },
-    { label: 'comedy', value: 'comedy' },
-    { label: 'drama', value: 'drama' },
-    { label: 'adventure', value: 'adventure' },
-    { label: 'crime', value: 'crime' },
-];
-
-// style for react-select component
-const selectStyles = {
-    container: (provided) => ({
-        ...provided,
-        width: '60%', 
-        minWidth: '300px',
-        textAlign: "center",
-    }),
-    control: (provided) => ({
-        ...provided,
-        borderRadius: "10px",
-        width: '100%',
-    }),
-};
 // overriding leftSide style
 const leftSideStyle = {
     display: 'flex',
@@ -202,28 +165,7 @@ function SignupPersonalize() {
                                 />
                             </div>
                             <div className="genre-select">
-                                <h4>Genre Preferences</h4>
-                                <Select 
-                                    className='dropdown' 
-                                    options={genreOptions} 
-                                    onChange={opt => setGenrePreferences(opt)}
-                                    isMulti
-                                    menuPlacement="auto"
-                                    styles={selectStyles}
-                                    theme={(theme) => ({
-                                        ...theme,
-                                        borderRadius: 0,
-                                        colors: {
-                                            ...theme.colors,
-                                            /* when hovering */
-                                            primary25: '#97AD97',
-                                            /* when clicking */
-                                            primary50: '#97AD97',
-                                            /* border color of the select dropbox */
-                                            primary: '#F0ABFB',
-                                        },
-                                        })}
-                                />
+                                <GenrePreferences return_genres={setGenrePreferences}/>
                             </div>
                         </div>
                     </div>
